Svg rendering

+++ MarkedMap.cpp
@@ -6,27 +6,25 @@ MarkedMap::MarkedMap(const std::string &filename) : Map(filename)
 
 std::pair<int, int> MarkedMap::getHeroPosition() const
 {
-	for (int i = 0; i < map.size(); i++)
+	for (size_t i = 0; i < map.size(); i++)
 	{
-		for (int j = 0; j < map[i].size(); j++)
+		for (size_t j = 0; j < map[i].size(); j++)
 		{
 			if (map[i][j] == 'H')
 			{
 				return std::pair<int, int>(i, j);
 			}
-			
 		}
 	}
 	throw WrongMapTypeException();
 }
 
-
 std::vector<std::pair<int, int>> MarkedMap::getMonsterPositions(char c) const
 {
 	std::vector<std::pair<int, int>> positions;
-	for (int i = 0; i < map.size(); i++)
+	for (size_t i = 0; i < map.size(); i++)
 	{
-		for (int j = 0; j < map[i].size(); j++)
+		for (size_t j = 0; j < map[i].size(); j++)
 		{
 			if (map[i][j] == c)
 			{
@@ -38,4 +36,3 @@ std::vector<std::pair<int, int>> MarkedMap::getMonsterPositions(char c) const
 		throw WrongMapTypeException();
 	return positions;
 }
-
